[snapshot] Fix builtin's FeedbackCellArray corrupted by serializer
This CL fixes a crash which started happening after landing
https://chromium-review.googlesource.com/c/v8/v8/+/2345966.

It causes closure feedback cell arrays owned by non-JS code objects (e.g
builtins) to be discarded, while the code object itself is left intact.

This is problematic because
CodeStubAssembler::LoadClosureFeedbackArray() expects to find either a
FeedbackVector or a FeedbackCellArray in the feedback cell. If it
encounters `undefined` there, this crash happens:

  Fatal error in v8/src/compiler/code-assembler.cc, line 1671
  Type cast failed in CAST(feedback_cell_array.value())
    at v8/src/codegen/code-stub-assembler.cc:9288
  Expected FeedbackVector but found
    0x24c608042309: [Oddball] in ReadOnlySpace: #undefined